requirements.txt
# ─── MaizeMedic — requirements.txt ────────────────────────────────────────────
# Python 3.10 recommended
# Streamlit Cloud: uses tensorflow-cpu (no GPU on free tier)
# Local GPU:       replace tensorflow-cpu with tensorflow

# ── Web framework ──────────────────────────────────────────────────────────streamlit>=1.32.0,<2.0.0

# ── Deep learning ──────────────────────────────────────────────────────────
tensorflow-cpu==2.15.0

# ── Image processing ──────────────────────────────────────────────────────────
Pillow>=10.2.0
opencv-python-headless>=4.9.0.80

# ── Numerical & visualisation ──────────────────────────────────────────────────
numpy>=1.26.0,<2.0.0
matplotlib>=3.8.0

# ── Chatbot backends ──────────────────────────────────────────────────────────
requests>=2.31.0     # Ollama (local)
groq>=0.4.0          # Groq API (Streamlit Cloud)

# ── Model auto-download from Google Drive ─────────────────────────────────────
gdown>=5.1.0